+++ Assets/ScaleMonk Ads/Source/AdsEditorBinding.cs
@@ -5,6 +5,7 @@
 // https://www.scalemonk.com/legal/en-US/mediation-license-agreement/index.html 
 //
 
+using System;
 using System.Collections.Generic;
 using System.Runtime.CompilerServices;
 using UnityEngine;
@@ -18,6 +19,8 @@ namespace ScaleMonk.Ads
         public void Initialize(ScaleMonkAds adsInstance)
         {
             _scaleMonkAds = adsInstance;
+            _scaleMonkAds.InitializationCompleted();
+            
             Debug.Log("ScaleMonkAds initialized successfully");
         }
         
@@ -46,14 +49,27 @@ namespace ScaleMonk.Ads
             mockAdInstance = GameObject.Instantiate(mockAdPrefab);
             
             var bannerCanvas = (RectTransform) mockAdInstance.transform.Find("Canvas/BgColor");
+
+            // Remove title from Rectangle banners because it's too small to show it.
+            if (bannerSize == BannerSize.Rectangle)
+            {
+                var textChild = bannerCanvas.transform.Find("Title");
+                textChild.parent = null;
+            }
+
+            ResizeAndPositionBannerCanvas(bannerSize, bannerPosition, bannerCanvas);
+
+            return mockAdInstance;
+        }
+
+        private static void ResizeAndPositionBannerCanvas(BannerSize bannerSize, BannerPosition bannerPosition,
+            RectTransform bannerCanvas)
+        {
             var editorPosition = bannerPosition.toEditorPosition();
-            
             bannerCanvas.sizeDelta = new Vector2(bannerSize.Width, bannerSize.Height);
             bannerCanvas.anchorMin = editorPosition.AnchorMin;
             bannerCanvas.anchorMax = editorPosition.AnchorMax;
             bannerCanvas.pivot = editorPosition.Pivot;
-
-            return mockAdInstance;
         }
 #endif
         public void ShowInterstitial(string tag)
@@ -80,8 +96,11 @@ namespace ScaleMonk.Ads
         {
             Debug.Log("Banner shown at " + tag);
 #if UNITY_2018_4_OR_NEWER
-            _scaleMonkAds.CompletedBannerDisplay(tag);
-            _banner = CreateBannerMockAdInstance(bannerSize, bannerPosition);
+            if (_banner == null)
+            {
+                _scaleMonkAds.CompletedBannerDisplay(tag);
+                _banner = CreateBannerMockAdInstance(bannerSize, bannerPosition);   
+            }
 #endif
         }